add labels for password and confirm
