+++ src/packages/auth-ui-components/src/components/password/confirm-password/index.ts
@@ -1,0 +1,1 @@
+export * from './component';
