+++ src/packages/auth-ui-components/src/components/password/confirm-password/styles.module.css
@@ -1,0 +1,22 @@
+.textInputField {
+	box-sizing: border-box;
+	height: 40px;
+	background: rgba(237, 232, 242, 0.02);
+	border: 1px solid rgba(237, 232, 242, 0.1);
+	border-radius: 6px;
+	padding: 10px;
+	width: 100%;
+}
+
+.fieldLabel {
+	flex: none;
+
+	font-family: 'Inter';
+	font-style: normal;
+	font-size: 14px;
+	line-height: 150%;
+	/* identical to box height, or 21px */
+
+	color: rgba(237, 232, 242, 0.6);
+	font-weight: 600;
+}
