+++ src/packages/auth-ui-components/src/components/password/index.ts
@@ -1,2 +1,3 @@
 export * from './challenge';
 export * from './login';
+export * from './confirm-password';
