+++ src/packages/auth-ui-components/src/components/password/login/component.tsx
@@ -13,35 +13,12 @@ import styles from './styles.module.css';
 
 import { LOGIN_MUTATION } from './graphql';
 import { formatRedirectUrl } from '../../../utils/urls';
+import { PasswordComponent } from '../confirm-password';
 
 interface Form {
 	username: string;
 	password: string;
 }
-
-export const PasswordComponent = () => {
-	return (
-		<Field
-			type="password"
-			placeholder="Password"
-			id="password"
-			name="password"
-			className={styles.textInputField}
-		/>
-	);
-};
-
-export const ConfirmComponent = () => {
-	return (
-		<Field
-			type="password"
-			placeholder="Confirm"
-			id="confirm"
-			name="confirm"
-			className={styles.textInputField}
-		/>
-	);
-};
 
 export const PasswordLogin = () => {
 	const [login] = useMutation<{ result: { authToken: string } }>(LOGIN_MUTATION);
